Move and refactor import analyzer
